clean: prefer unsigned ints for things that shouldn't be negative
tdd: refactor

+++ render/rendertest/glcontext.go
@@ -293,8 +293,6 @@ func newGlWindowForTest(width, height int) (system.System, system.NativeWindowHa
 	return sys, <-hdl, renderQueue
 }
 
-// TODO(clean): don't take a signed value; we don't want to imply we can rewind
-// time!
-func AdvanceTime(sys system.System, delta int64) {
-	sys.(*system.MockSystem).AdvanceTime(uint64(delta))
+func AdvanceTime(sys system.System, delta uint64) {
+	sys.(*system.MockSystem).AdvanceTime(delta)
 }
